Adjest proper MTU for bridges
If setting MTU on a bridge (bonds too but they adjust automatically),
it should not exceed the smallest MTU value of any ports in that bridge.
Since all of the ports had default MTU (1500) we were erroring out.
Changed so in one case we do not set MTU on a bridge and the other one
where we do, but also adjust ports.